refactor(forge-dsl): P0 机械清理——硬编码路径 / 重复函数 / 文档
- lib.rs：删除写死开发机路径的 dump（D:\Program\rust_main\code_forge\...，
  已有 FGE_DEBUG_GEN 机制的 dump_generated 覆盖）
- 新增 v12/shared.rs：parse_u64（原 codegen/mod.rs 与 validate.rs 各一份，
  完全一致）与 group_names（原 codegen/mod.rs 与 integration.rs 各一份）
  单点定义，两处删除本地副本改引用
- codegen/mod.rs 头部重复文档段合并

等价变换：forge-dsl lib 46、forge-codegen jit 全套无 FAILED。

+++ crates/backend/forge-codegen/src/machine/frame.rs
@@ -56,7 +56,7 @@ pub trait TargetFrameLowering: Send + Sync + 'static {
         &self,
         _dst_reg: u32,
         _offset: i32,
-        _width: u8,
+        _width: u16,
         _is_fp: bool,
         _sink: &mut CodeSink,
     ) -> Result<(), IrError> {
@@ -71,7 +71,7 @@ pub trait TargetFrameLowering: Send + Sync + 'static {
         &self,
         _src_reg: u32,
         _offset: i32,
-        _width: u8,
+        _width: u16,
         _is_fp: bool,
         _sink: &mut CodeSink,
     ) -> Result<(), IrError> {
